Changed the type of the phys_virt_offset to a uintptr_t.
This variable is the difference from the physical address to the
virtual address of a user application. It was a uint32_t for all
architectures. This meant that if you were running a 64-bit CPU
and the physical address was less than the virtual address, then
the phys_virt_offset would be truncated and the kernel would crash
during initialization of the User Application. Now we use a type
that is based on the bit size.

+++ elfloader-tool/include/arch-arm/elfloader.h
@@ -51,14 +51,14 @@ struct image_info {
      *
      *  virtual_address + phys_virt_offset = physical_address
      */
-    uint32_t phys_virt_offset;
+    uintptr_t phys_virt_offset;
 };
 
 extern struct image_info kernel_info;
 extern struct image_info user_info;
 typedef void (*init_kernel_t)(paddr_t ui_p_reg_start,
                               paddr_t ui_p_reg_end,
-                              int32_t pv_offset,
+                              uintptr_t pv_offset,
                               vaddr_t v_entry);
 
 /* Enable the mmu. */
